+++ lib/cinder/components/filters.ex
@@ -38,6 +38,7 @@ defmodule Cinder.Components.Filters do
     :filter_select_option_class,
     :filter_select_label_class,
     :filter_select_empty_class,
+    :filter_select_show_arrow,
 
     # Boolean filter
     :filter_boolean_container_class,
@@ -107,6 +108,7 @@ defmodule Cinder.Components.Filters do
       filter_select_option_class: "",
       filter_select_label_class: "",
       filter_select_empty_class: "",
+      filter_select_show_arrow: true,
 
       # Boolean filter
       filter_boolean_container_class: "",