+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Bug fixes
 
 * Merge provided `filter_options` with default options for a column, instead of overwriting them
+* Fix slight input jumping issues across all themes and duplicate select arrows from DaisyUI theme
 
 ### Chores